trabalho2/requirements.txt
glfw==2.6.2
PyOpenGL==3.1.7
numpy==1.23.3
Pillow==9.2.0
PyGLM==2.7.0
termcolor==2.3.0
